+++ src/components/App.jsx
@@ -2,10 +2,10 @@ import React from 'react';
 
 import { Link, Route, Switch } from 'react-router-dom';
 
-import HomePage from './pages/Home/HomePage';
-import AboutPage from './pages/About/AboutPage';
-import RestaurantsPage from './pages/Restaurants/RestaurantsPage';
-import NotFoundPage from './pages/NotFound/NotFoundPage';
+import HomePage from '../pages/Home/HomePage';
+import AboutPage from '../pages/About/AboutPage';
+import RestaurantsPage from '../pages/Restaurants/RestaurantsPage';
+import NotFoundPage from '../pages/NotFound/NotFoundPage';
 
 export default function App() {
   return (
